Added try/catches for null edits/deletes, scaled UI for laptop
Plan on presenting this on laptop, even on desktop it'll just appear to
be larger than it probably should be. Try/catches added to handle
NullPointerException as a result of trying to grab currentlySelected
when there isn't anything there - found as a result of playing with the
edit/delete buttons while nothing was loaded.